requirements.txt
openpyxl
pydantic
pillow
pymupdf
openai
gradio
pathlib
